pbs-156-feat: user token 검증 -  remain domain
# Conflicts:
#	backend/src/main/java/com/pobluesky/backend/domain/answer/service/AnswerService.java
#	backend/src/main/java/com/pobluesky/backend/domain/inquiry/repository/InquiryRepository.java
#	backend/src/main/java/com/pobluesky/backend/domain/inquiry/service/InquiryService.java
#	backend/src/main/java/com/pobluesky/backend/domain/lineitem/service/LineItemService.java
#	backend/src/main/java/com/pobluesky/backend/domain/question/repository/QuestionRepository.java
#	backend/src/main/java/com/pobluesky/backend/domain/question/service/QuestionService.java
#	backend/src/main/java/com/pobluesky/backend/global/security/JwtTokenProvider.java
#	backend/src/main/resources/application.properties
#	backend/src/main/resources/data.sql

+++ backend/src/main/java/com/pobluesky/backend/domain/inquiry/service/InquiryService.java
@@ -7,16 +7,22 @@ import com.pobluesky.backend.domain.inquiry.entity.Inquiry;
 import com.pobluesky.backend.domain.inquiry.entity.Progress;
 import com.pobluesky.backend.domain.inquiry.repository.InquiryRepository;
 import com.pobluesky.backend.domain.user.entity.Customer;
+import com.pobluesky.backend.domain.user.entity.Manager;
+import com.pobluesky.backend.domain.user.entity.UserRole;
+
 import com.pobluesky.backend.domain.user.repository.CustomerRepository;
-import com.pobluesky.backend.domain.user.service.CustomerService;
+import com.pobluesky.backend.domain.user.repository.ManagerRepository;
+import com.pobluesky.backend.domain.user.service.SignService;
 import com.pobluesky.backend.global.error.CommonException;
 import com.pobluesky.backend.global.error.ErrorCode;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.relational.core.sql.In;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,12 +30,27 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 @RequiredArgsConstructor
 public class InquiryService {
+
+    private final SignService signService;
+
     private final InquiryRepository inquiryRepository;
+
     private final CustomerRepository customerRepository;
 
+    private final ManagerRepository managerRepository;
+
     @Transactional(readOnly = true)
-    public List<InquiryResponseDTO> getInquiriesByCustomerId(Long customerId) {
-        List<Inquiry> inquiries = inquiryRepository.findByCustomer_CustomerIdAndIsActivatedTrue(customerId);
+    public List<InquiryResponseDTO> getInquiriesByCustomerId(String token, Long customerId) {
+        Long userId = signService.parseToken(token);
+
+        Customer customer = customerRepository.findById(userId)
+            .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
+
+        if(!Objects.equals(customer.getUserId(), customerId))
+            throw new CommonException(ErrorCode.USER_NOT_MATCHED);
+
+        List<Inquiry> inquiries =
+            inquiryRepository.findByCustomer_UserIdAndIsActivatedTrue(customerId);
 
         return inquiries.stream()
             .map(InquiryResponseDTO::from)
@@ -37,7 +58,15 @@ public class InquiryService {
     }
 
     @Transactional
-    public List<InquiryResponseDTO> getInquiries() {
+    public List<InquiryResponseDTO> getInquiries(String token) {
+        Long userId = signService.parseToken(token);
+
+        Manager user =managerRepository.findById(userId)
+            .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
+
+        if(user.getRole() == UserRole.CUSTOMER)
+            throw new CommonException(ErrorCode.UNAUTHORIZED_USER_MANAGER);
+
         List<Inquiry> inquiries = inquiryRepository.findAll();
 
         return inquiries.stream()
@@ -46,9 +75,18 @@ public class InquiryService {
     }
 
     @Transactional
-    public InquiryResponseDTO createInquiry(Long customerId,InquiryCreateRequestDTO dto) {
-        Customer customer = customerRepository.findById(customerId)
+    public InquiryResponseDTO createInquiry(
+        String token,
+        Long customerId,
+        InquiryCreateRequestDTO dto
+    ) {
+        Long userId = signService.parseToken(token);
+
+        Customer customer = customerRepository.findById(userId)
             .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
+
+        if(!Objects.equals(customer.getUserId(), customerId))
+            throw new CommonException(ErrorCode.USER_NOT_MATCHED);
 
         Inquiry inquiry = dto.toInquiryEntity();
         inquiry.setCustomer(customer);
@@ -58,11 +96,21 @@ public class InquiryService {
     }
 
     @Transactional
-    public InquiryResponseDTO updateInquiryById(Long inquiryId, InquiryUpdateRequestDTO inquiryUpdateRequestDTO) {
+    public InquiryResponseDTO updateInquiryById(
+        String token,
+        Long inquiryId,
+        InquiryUpdateRequestDTO inquiryUpdateRequestDTO
+    ) {
+        Long userId = signService.parseToken(token);
+
+        Customer customer = customerRepository.findById(userId)
+            .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
 
         Inquiry inquiry = inquiryRepository.findById(inquiryId)
             .orElseThrow(() -> new CommonException(ErrorCode.INQUIRY_NOT_FOUND));
 
+        if(!Objects.equals(customer.getUserId(), inquiry.getCustomer().getUserId()))
+            throw new CommonException(ErrorCode.USER_NOT_MATCHED);
 
         inquiry.updateInquiry(
             inquiryUpdateRequestDTO.country(),
@@ -87,31 +135,61 @@ public class InquiryService {
     }
 
     @Transactional
-    public void deleteInquiryById(Long inquiryId) {
+    public void deleteInquiryById(String token, Long inquiryId) {
+        Long userId = signService.parseToken(token);
+
+        Customer customer = customerRepository.findById(userId)
+            .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
 
         Inquiry inquiry = inquiryRepository.findById(inquiryId)
             .orElseThrow(() -> new CommonException(ErrorCode.INQUIRY_NOT_FOUND));
 
-        inquiry.deleteInquiry();
+        if (!Objects.equals(customer.getUserId(), inquiry.getCustomer().getUserId()))
+            throw new CommonException(ErrorCode.USER_NOT_MATCHED);
 
+        inquiry.deleteInquiry();
     }
 
     @Transactional(readOnly = true)
-    public InquiryResponseDTO getInquiryDetail(Long customerId, Long inquiryId) {
-        Customer customer = customerRepository.findById(customerId)
+    public InquiryResponseDTO getInquiryDetail(
+        String token,
+        Long customerId,
+        Long inquiryId
+    ) {
+        Long userId = signService.parseToken(token);
+
+        Customer customer = customerRepository.findById(userId)
             .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
-        Inquiry inquiry = (Inquiry) inquiryRepository.findByCustomer_CustomerIdAndInquiryId(customerId,inquiryId)
+
+        inquiryRepository.findById(inquiryId)
             .orElseThrow(() -> new CommonException(ErrorCode.INQUIRY_NOT_FOUND));
+
+        Inquiry inquiry = inquiryRepository.findByCustomer_UserIdAndInquiryId(customerId,inquiryId)
+            .orElseThrow(() -> new CommonException(ErrorCode.INQUIRY_NOT_FOUND));
+
+        if(!Objects.equals(customer.getUserId(), inquiry.getCustomer().getUserId()))
+            throw new CommonException(ErrorCode.USER_NOT_MATCHED);
 
         return InquiryResponseDTO.from(inquiry);
     }
 
     @Transactional(readOnly = true)
-    public List<InquiryResponseDTO> getInquiriesByProgress(Progress progress) {
+    public List<InquiryResponseDTO> getInquiriesByProgress(String token, Progress progress) {
+        Long userId = signService.parseToken(token);
+
+        Stream.of(
+                customerRepository.findById(userId),
+                managerRepository.findById(userId)
+            )
+            .flatMap(Optional::stream)
+            .findFirst()
+            .orElseThrow(() -> new CommonException(ErrorCode.USER_NOT_FOUND));
+
         List<Inquiry> inquiries = inquiryRepository.findByProgress(progress);
 
         return inquiries.stream()
             .map(InquiryResponseDTO::from)
             .collect(Collectors.toList());
     }
+
 }
